Move admin pack routes to blueprint
