+++ app.py
@@ -155,39 +155,6 @@ def get_all_comments():
     return _cmt_mgr().get_all_comments()
 
 
-# ----- Admin Packs CRUD -----
-from modules.admin import (
-    admin_packs_list,
-    admin_packs_new,
-    admin_packs_edit,
-    admin_packs_delete,
-)
-from utils.security import admin_required
-
-
-@admin_bp.route('/packs')
-@admin_required
-def admin_packs():
-    return admin_packs_list()
-
-
-@admin_bp.route('/packs/new', methods=['GET', 'POST'])
-@admin_required
-def admin_packs_new_route():
-    return admin_packs_new()
-
-
-@admin_bp.route('/packs/<string:slug>/edit', methods=['GET', 'POST'])
-@admin_required
-def admin_packs_edit_route(slug):
-    return admin_packs_edit(slug)
-
-
-@admin_bp.route('/packs/<string:slug>/delete', methods=['POST'])
-@admin_required
-def admin_packs_delete_route(slug):
-    return admin_packs_delete(slug)
-
 # ------------- Forum routes (remain here) -------------
 
 @app.route('/forum')
